first iteration
I have used the youtube api to pull the latest 5 videos from the TED youtube channel.
I ahve styled the page so it changes how the videos are layed out. there is a mobile friendly version then a smaller screen version and then a +1000 pixels version.
future updates:
review js code, make sure its the most effecient
make the call to the youtube api once instead of making it call everytime the page is loaded (just store in local data and check it when the page is loaded, if empty them make the call????)